[refactor] 채널의 store 내부에 state명(showList -> showChannelList) 변경으로 파일 수정
