+++ client/src/components/ChannelListBox/ChannelList/Channel/ChannelItem.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useDispatch } from 'react-redux';
-import { onChangeCurrent } from '@/store/modules/channels';
+import { onChangeCurrent, loadChannelRequest } from '@/store/modules/channel';
 import styled from 'styled-components';
-import { useChannelList, useChannels } from '@/hooks/useChannels';
+import { useChannelList, useChannel } from '@/hooks/useChannel';
 
 interface Args {
   key: number;
@@ -21,7 +21,7 @@ const ChannelWrapper = styled.div<Props>`
   font-size: 20px;
   color: #fff;
   &: hover {
-    background: rgba(0, 0, 0, 0.2);
+    ${(props: Props) => (!props.background ? 'background: rgba(0, 0, 0, 0.2);' : '')}
   }
   background: ${(props: Props) => (props.background ? 'blue' : 'transparent')};
 `;
@@ -35,12 +35,17 @@ const NameWrapper = styled.div``;
 const ChannelItem = (args: Args) => {
   const dispatch = useDispatch();
   const { id, name, isPublic } = useChannelList(args.idx);
-  const { current } = useChannels();
+  const { current } = useChannel();
 
-  const onClick = () => dispatch(onChangeCurrent(id));
+  const onClick = (e: React.MouseEvent<HTMLElement>) => {
+    e.preventDefault();
+
+    dispatch(onChangeCurrent(args.idx));
+    dispatch(loadChannelRequest(id));
+  };
 
   return (
-    <ChannelWrapper onClick={onClick} role="button" background={id === current}>
+    <ChannelWrapper onClick={onClick} role="button" background={id === current?.id}>
       <IconWrapper>{isPublic === 1 ? '#' : 'O'}</IconWrapper>
       <NameWrapper>{name}</NameWrapper>
     </ChannelWrapper>
